Class `RepoIdentity` moved to module `repo_identity.py`
