+++ commitfetch/commit.py
@@ -1,16 +1,17 @@
 from datetime import datetime
 from pathlib import Path
 
+from . import recurring_strings as rs
+from .repo_identity import RepoIdentity
+
 
 _DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%SZ"
 
-_CLOSING_PAR = ")"
-_COMMA_SPACE = ", "
-_OPENING_PAR = "("
-_QUOTE = "'"
-_QUOTE_CLOSING_PAR = _QUOTE + _CLOSING_PAR
-_QUOTE_COMMA_SPACE = _QUOTE + _COMMA_SPACE
-_SLASH = "/"
+def _datetime_from_str(datetime_str):
+	return datetime.strptime(datetime_str, _DATETIME_FORMAT)
+
+def _datetime_to_str(datetime_obj):
+	return datetime.strftime(datetime_obj, _DATETIME_FORMAT)
 
 
 class Commit:
@@ -58,13 +59,13 @@ class Commit:
 	def __repr__(self):
 		str_paths = *(str(file) for file in self._files),
 
-		return self.__class__.__name__ + _OPENING_PAR\
-			+ _QUOTE + self._sha + _QUOTE_COMMA_SPACE\
-			+ repr(self._message) + _COMMA_SPACE\
-			+ _QUOTE + str(self._repository) + _QUOTE_COMMA_SPACE\
-			+ _QUOTE + self._author + _QUOTE_COMMA_SPACE\
-			+ _QUOTE + self.moment_to_str() + _QUOTE_COMMA_SPACE\
-			+ str(str_paths) + _CLOSING_PAR
+		return self.__class__.__name__ + rs.OPENING_PAR\
+			+ rs.QUOTE + self._sha + rs.QUOTE_COMMA_SPACE\
+			+ repr(self._message) + rs.COMMA_SPACE\
+			+ rs.QUOTE + str(self._repository) + rs.QUOTE_COMMA_SPACE\
+			+ rs.QUOTE + self._author + rs.QUOTE_COMMA_SPACE\
+			+ rs.QUOTE + self.moment_to_str() + rs.QUOTE_COMMA_SPACE\
+			+ str(str_paths) + rs.CLOSING_PAR
 
 	@property
 	def author(self):
@@ -118,90 +119,3 @@ class Commit:
 		str: the SHA hash that identifies this commit.
 		"""
 		return self._sha
-
-
-class RepoIdentity:
-	"""
-	This class groups the name of a GitHub repository's owner and the
-	repository's name, which constitute the repository's identity.
-	"""
-
-	def __init__(self, owner, name):
-		"""
-		The constructor requires the repository's owner and name.
-
-		Parameters:
-			owner (str): the name of the repository's owner.
-			name (str): the repository's name.
-		"""
-		self._owner = owner
-		self._name = name
-
-	def __repr__(self):
-		return self.__class__.__name__ + _OPENING_PAR\
-			+ _QUOTE + self._owner + _QUOTE_COMMA_SPACE\
-			+ _QUOTE + self._name + _QUOTE_CLOSING_PAR\
-
-	def __str__(self):
-		return self.get_full_name()
-
-	@staticmethod
-	def from_full_name(full_name):
-		"""
-		Makes a RepoIdentity from a repository's full name, in the format
-		<owner>/<name>.
-
-		Parameters:
-			full_name (str): a repository's full name in the prescibed format.
-
-		Returns:
-			RepoIdentity: an object that identifies the indicated repository.
-
-		Raises:
-			ValueError: if parameter full_name is not in the expected format.
-		"""
-		split_name = full_name.split(_SLASH)
-
-		if len(split_name) != 2:
-			raise ValueError(
-				"The repository name must be in the format <owner>/<name>.")
-
-		return RepoIdentity(split_name[0], split_name[1])
-
-	def get_full_name(self, separator=_SLASH):
-		"""
-		Provides the repository's full name by joining the owner's name and the
-		repository's name with the given separator. Calling this method without
-		specifying a separator is equivalent to using function str.
-
-		Parameters:
-			separator (str): It comes between the owner's name and the
-				repository's name. Defaults to '/'.
-
-		Returns:
-			str: the repository's full name in the format
-				<owner><separator><name>.
-		"""
-		return self._owner + separator + self._name
-
-	@property
-	def name(self):
-		"""
-		str: the repository's name.
-		"""
-		return self._name
-
-	@property
-	def owner(self):
-		"""
-		str: the name of the repository's owner.
-		"""
-		return self._owner
-
-
-def _datetime_from_str(datetime_str):
-	return datetime.strptime(datetime_str, _DATETIME_FORMAT)
-
-
-def _datetime_to_str(datetime_obj):
-	return datetime.strftime(datetime_obj, _DATETIME_FORMAT)
