+++ commitfetch/recurring_strings.py
@@ -1,0 +1,6 @@
+CLOSING_PAR = ")"
+COMMA_SPACE = ", "
+OPENING_PAR = "("
+QUOTE = "'"
+QUOTE_CLOSING_PAR = QUOTE + CLOSING_PAR
+QUOTE_COMMA_SPACE = QUOTE + COMMA_SPACE
